+++ src/features/statusBar/statusBar.ts
@@ -9,6 +9,7 @@ export class StatusBar {
   private languageSupport: string[] = [];
   private hoverEnabled: boolean = true;
   private colorPreviewEnabled: boolean = true;
+  private sortOnSave: boolean = false;
   private callbacks: ((
     isActive: boolean,
     useLocalFile: boolean,
@@ -38,6 +39,7 @@ export class StatusBar {
         languageSupport: string[];
         enableHover: boolean;
         enableColorPreview: boolean;
+        sortOnSave: boolean;
       }>('bootstrapIntelliSense');
 
       this.isActive = bootstrapConfig?.enable ?? true;
@@ -47,6 +49,7 @@ export class StatusBar {
       this.languageSupport = bootstrapConfig?.languageSupport ?? [];
       this.hoverEnabled = bootstrapConfig?.enableHover ?? true;
       this.colorPreviewEnabled = bootstrapConfig?.enableColorPreview ?? true;
+      this.sortOnSave = bootstrapConfig?.sortOnSave ?? false;
     } catch (error) {
       this.isActive = true;
       this.bootstrapVersion = this.defaultVersion;
@@ -55,6 +58,7 @@ export class StatusBar {
       this.languageSupport = [];
       this.hoverEnabled = true;
       this.colorPreviewEnabled = true;
+      this.sortOnSave = false;
     }
   }
 
@@ -85,6 +89,7 @@ export class StatusBar {
         languageSupport: this.languageSupport,
         enableHover: this.hoverEnabled,
         enableColorPreview: this.colorPreviewEnabled,
+        sortOnSave: this.sortOnSave,
       };
       await config.update('bootstrapIntelliSense', settings, this.getConfigurationTarget(config));
     } catch (error) {
@@ -146,6 +151,28 @@ export class StatusBar {
 
   public getColorPreviewEnabled(): boolean {
     return this.colorPreviewEnabled;
+  }
+
+  public getSortOnSave(): boolean {
+    return this.sortOnSave;
+  }
+
+  public async toggleSortOnSave() {
+    const oldStatus = this.sortOnSave;
+    this.sortOnSave = !this.sortOnSave;
+
+    try {
+      await this.saveSettings();
+      this.callbacks.forEach((callback) =>
+        callback(this.isActive, this.useLocalFile, this.cssFilePath, this.bootstrapVersion, this.languageSupport),
+      );
+
+      const statusChange = this.sortOnSave ? 'enabled' : 'disabled';
+      vscode.window.showInformationMessage(`Bootstrap IntelliSense class sorting on save has been ${statusChange}`);
+    } catch (error) {
+      this.sortOnSave = oldStatus;
+      vscode.window.showErrorMessage('Error toggling Bootstrap IntelliSense class sorting');
+    }
   }
 
   public async toggleColorPreview() {